backend/pyproject.toml
[project]
name = "crm-backend"
version = "0.1.0"
description = "Local monolithic backend for CRM Workroom"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0,<4.0.0",
    "alembic>=1.14.0,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "PyJWT>=2.10.0,<3.0.0",
    "redis>=5.2.0,<6.0.0",
    "sqlalchemy>=2.0.38,<3.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.25.0,<1.0.0",
    "ruff>=0.9.0,<1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
